breaking down the main file into several files
